feat: Refactor message payloads and handlers to use camelCase for consistency and improve JSON serialization

+++ user-service/Infrastructure/Services/RealtimeNotificationService.cs
@@ -57,7 +57,7 @@ public class RealtimeNotificationService : INotificationService
                 message.Text,
                 message.UserId,
                 message.CreatedAt,
-                message.Status,
+                (int)message.Status,
                 message.Attachments ?? new List<MessageAttachmentDto>()
             )
         };
